code cleanup

+++ source/modules/joystick/JoystickModule.cpp
@@ -54,7 +54,7 @@ namespace love
 
     JoystickBase* JoystickModule::addJoystick(int64_t deviceId)
     {
-        if (deviceId < 0 || (size_t)deviceId >= joystick::getJoystickCount())
+        if (deviceId < 0 || (int)deviceId >= joystick::getJoystickCount())
             return nullptr;
 
         std::string guid       = this->getDeviceGUID(deviceId);
@@ -71,8 +71,6 @@ namespace love
             }
         }
 
-        // TODO: make a method (or use define info)
-        // for when Wii U handles are dealt with (WPADHandle, VPADHandle classes)
         if (!joystick)
         {
             joystick = love::joystick::openJoystick(this->joysticks.size());
